virt-common: init.virt.rc: Override bt_socket service to force disable it
android.hardware.bluetooth.socket's VINTF entry can't appear in the build for now because of:

INFO: Removing HAL from the manifest because it is declaring V1 of a new unfrozen interface which is not allowed in this release configuration: android.hardware.bluetooth.socket

therefore it doesn't start and keeps restarting all the times. Force disable it for now.

Change-Id: I6995cb0ba24e3fcb234f66335b2b79d566e65e18